Dispositivos y apps, y espacio de trabajo: pantallas reales
Las dos eran maqueta. "Dispositivos y apps" caía en el renderizador
genérico, con interruptores que se activaban según la posición de la fila;
el espacio de trabajo tenía el nombre y el ID escritos a mano.

Dispositivos y apps
- Sesiones reales desde auth.sessions vía función acotada al propio usuario:
  navegador, sistema, IP y última actividad, con "este dispositivo" marcado
- Cerrar una sesión concreta o todas
- Claves de API: se crean con token de un solo uso y en la base queda solo
  el hash SHA-256 y el prefijo; se pueden revocar

Espacio de trabajo
- Tabla workspaces con RLS; cada perfil estrena el suyo por trigger
- Nombre editable con contador, color de avatar, identificador público único
  y límite de créditos por miembro, todo persistido
- ID copiable real y recuento de proyectos
- Eliminar espacio con confirmación escrita
- El nombre real aparece en el sidebar y en la navegación de ajustes, que
  antes mostraban un nombre fijo

+++ backend/app/providers/hailuo.py
@@ -16,7 +16,6 @@ MiniMax factura por vídeo y no por segundo (informe 06 §2.1, cifras [S]), así
 
 from __future__ import annotations
 
-from decimal import Decimal
 from typing import Any
 
 from app.config import get_settings
@@ -64,7 +63,7 @@ class HailuoAdapter(HttpAdapter):
         payload: dict[str, Any] = {
             "model": _MODEL_NAME.get(req.model_id, req.model_id),
             "prompt": self._styled_prompt(req),
-            "duration": int(round(req.duration_s or 6)),
+            "duration": round(req.duration_s or 6),
         }
         if req.resolution:
             payload["resolution"] = req.resolution.upper()
